+++ packages/react/src/blocks/Symbol.tsx
@@ -112,8 +112,8 @@ class SymbolComponent extends React.Component<PropsWithChildren<SymbolProps>> {
       showPlaceholder = false;
     }
 
-    let key = dynamic ? undefined : [model, entry].join(':');
-    const dataString = Builder.isEditing ? null : data && size(data) && hash(data);
+    let key = dynamic ? this.props.builderBlock?.id : [model, entry].join(':');
+    const dataString = data && size(data) && hash(data);
 
     if (key && dataString && dataString.length < 300) {
       key += ':' + dataString;