+++ packages/sdks-tests/src/e2e-tests/state-binding.spec.ts
@@ -9,7 +9,7 @@ test.describe('State binding', () => {
         'Angular Gen2 event binding not working properly for repeat blocks.'
       );
       // hydration errors
-      test.fail(packageName === 'gen1-next14-pages' || packageName === 'gen1-remix');
+      test.fail(packageName === 'gen1-next14-pages');
 
       // flaky, can't `test.fail()`
       test.skip(