+++ packages/core/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @builder.io/sdk
 
+## 6.0.2
+
+### Patch Changes
+
+- 306f8d5: Fix: add missing `folded` and `keysHelperText` types to custom component `Input`
+- 306f8d5: Types: add `firstPublished` to BuilderContent
+
+## 6.0.1
+
+### Patch Changes
+
+- c822422: Fix: symbols will now show published content instead of preview/autosave content while editing a page
+
 ## 6.0.0
 
 ### Major Changes
@@ -9,7 +22,7 @@
 
 ### Patch Changes
 
-- 06b1124: Fix: Content API invocations will have `includeRefs` set to `true` by default.
+- 06b1124: Fix: remove `enrich=true` default option passed to the API and instead use `includeRefs=true` as default
 - 409aec9: Feat: add `meta` type to custom components
 - 40d572d: Renders Symbol correctly when apiEndpoint is 'content'
 - 2fc9fc5: Fix: `onChange` functions passed to builder inputs can now receive async functions